Tests/LibWeb/Text/expected/css/white-space.txt

normal
normal
preserve discard-inner
